requirements.txt
discord.py
firebase-admin
google-auth
python-dotenv